Moved core test to the tornaodweb test module.
Refs: #109
